Primera refactorización dle aplicativo
